// du-tu-pulmo_be/src/modules/common/enums/screening-status.enum.ts
export enum ScreeningStatusEnum {
  UPLOADED = 'UPLOADED',
  PENDING_AI = 'PENDING_AI',
  AI_PROCESSING = 'AI_PROCESSING',
  AI_COMPLETED = 'AI_COMPLETED',
  AI_FAILED = 'AI_FAILED',
  PENDING_DOCTOR = 'PENDING_DOCTOR',
  DOCTOR_REVIEWING = 'DOCTOR_REVIEWING',
  DOCTOR_COMPLETED = 'DOCTOR_COMPLETED',
  CANCELLED = 'CANCELLED',
}